inital checkin
