+++ src/middleware/jira-symmetric-jwt-middleware.test.ts
@@ -1,8 +1,6 @@
 import { createQueryStringHash, encodeSymmetric } from "atlassian-jwt";
 import express, { Application, NextFunction, Request, Response } from "express";
-import { when } from "jest-when";
 import supertest from "supertest";
-import { booleanFlag, BooleanFlags } from "~/src/config/feature-flags";
 import { getLogger } from "~/src/config/logger";
 import { jiraSymmetricJwtMiddleware } from "~/src/middleware/jira-symmetric-jwt-middleware";
 import { Installation } from "~/src/models/installation";
@@ -25,10 +23,6 @@ describe("jiraSymmetricJwtMiddleware", () => {
 			host: jiraHost,
 			sharedSecret: testSharedSecret
 		});
-
-		when(booleanFlag).calledWith(
-			BooleanFlags.NEW_JWT_VALIDATION
-		).mockResolvedValue(true);
 	});
 
 	it("should throw error when token is missing and no jiraHost in session", async () => {
