+++ src/routes/jira/configuration/jira-configuration-router.ts
@@ -2,11 +2,10 @@ import { Router } from "express";
 import { csrfMiddleware } from "middleware/csrf-middleware";
 import { JiraGet } from "routes/jira/jira-get";
 import { JiraDelete } from "routes/jira/jira-delete";
-import { JiraContextJwtTokenMiddleware, JiraJwtTokenMiddleware } from "~/src/middleware/jira-jwt-middleware";
 import { jiraSymmetricJwtMiddleware } from "~/src/middleware/jira-symmetric-jwt-middleware";
 
 export const JiraConfigurationRouter = Router();
 
 JiraConfigurationRouter.route("/")
-	.get(csrfMiddleware, JiraJwtTokenMiddleware, jiraSymmetricJwtMiddleware, JiraGet)
-	.delete(JiraContextJwtTokenMiddleware, jiraSymmetricJwtMiddleware, JiraDelete);
+	.get(csrfMiddleware, jiraSymmetricJwtMiddleware, JiraGet)
+	.delete(jiraSymmetricJwtMiddleware, JiraDelete);
