+++ src/routes/jira/connect/jira-connect-router.ts
@@ -2,12 +2,11 @@ import { Router } from "express";
 import { csrfMiddleware } from "middleware/csrf-middleware";
 import { JiraConnectGet } from "./jira-connect-get";
 import { JiraConnectEnterpriseRouter } from "./enterprise/jira-connect-enterprise-router";
-import { JiraJwtTokenMiddleware } from "~/src/middleware/jira-jwt-middleware";
 import { jiraSymmetricJwtMiddleware } from "~/src/middleware/jira-symmetric-jwt-middleware";
 
 export const JiraConnectRouter = Router();
 
 JiraConnectRouter.route("/")
-	.get(csrfMiddleware, JiraJwtTokenMiddleware, jiraSymmetricJwtMiddleware, JiraConnectGet);
+	.get(csrfMiddleware, jiraSymmetricJwtMiddleware, JiraConnectGet);
 
 JiraConnectRouter.use("/enterprise", JiraConnectEnterpriseRouter);
