+++ src/routes/jira/jira-router.ts
@@ -9,13 +9,12 @@ import { JiraDelete } from "routes/jira/jira-delete";
 import { JiraConnectRouter } from "routes/jira/connect/jira-connect-router";
 import { body } from "express-validator";
 import { returnOnValidationError } from "routes/api/api-utils";
-import { JiraContextJwtTokenMiddleware, JiraJwtTokenMiddleware } from "~/src/middleware/jira-jwt-middleware";
 import { jiraSymmetricJwtMiddleware } from "~/src/middleware/jira-symmetric-jwt-middleware";
 
 export const JiraRouter = Router();
 
 // TODO: The params `installationId` needs to be replaced by `subscriptionId`
-JiraRouter.delete("/subscription/:installationId", JiraContextJwtTokenMiddleware, jiraSymmetricJwtMiddleware, JiraDelete);
+JiraRouter.delete("/subscription/:installationId", jiraSymmetricJwtMiddleware, JiraDelete);
 
 JiraRouter.get("/atlassian-connect.json", JiraAtlassianConnectGet);
 
@@ -24,13 +23,12 @@ JiraRouter.use("/connect", JiraConnectRouter);
 JiraRouter.post("/sync",
 	body("commitsFromDate").optional().isISO8601(),
 	returnOnValidationError,
-	JiraContextJwtTokenMiddleware,
 	jiraSymmetricJwtMiddleware,
 	JiraSyncPost);
 
 JiraRouter.use("/events", JiraEventsRouter);
 
-JiraRouter.get("/", csrfMiddleware, JiraJwtTokenMiddleware, jiraSymmetricJwtMiddleware, JiraGet);
+JiraRouter.get("/", csrfMiddleware, jiraSymmetricJwtMiddleware, JiraGet);
 
 /********************************************************************************************************************
  * TODO: remove this later, keeping this for now cause its out in `Prod`
