+++ src/routes/router.ts
@@ -16,7 +16,6 @@ import { PublicRouter } from "./public/public-router";
 import { createAppClient } from "~/src/util/get-github-client-config";
 import { GithubManifestGet } from "routes/github/manifest/github-manifest-get";
 import { GithubCreateBranchOptionsGet } from "~/src/routes/github/create-branch/github-create-branch-options-get";
-import { jirahostMiddleware } from "~/src/middleware/jirahost-middleware";
 import { jiraSymmetricJwtMiddleware } from "~/src/middleware/jira-symmetric-jwt-middleware";
 
 export const RootRouter = Router();
@@ -57,9 +56,6 @@ RootRouter.use(MaintenanceRouter);
 RootRouter.get(["/session", "/session/*"], SessionGet);
 
 RootRouter.use(cookieSessionMiddleware);
-
-// Saves the jiraHost cookie to the secure session if available
-RootRouter.use(jirahostMiddleware);
 
 // App Manifest flow route
 RootRouter.get("/github-manifest", GithubManifestGet);
