+++ static/js/jira-configuration.js
@@ -19,8 +19,6 @@ $(".add-organization-link").click(function(event) {
 	const queryParameter = $(this).data("gh-cloud") ? "" : "?ghRedirect=to";
 	AP.context.getToken(function(token) {
 		const child = openChildWindow("/session/github/configuration" + queryParameter);
-		// Remove below line on cleaning up NEW_JWT_VALIDATION flag
-		child.window.jiraHost = jiraHost;
 		child.window.jwt = token;
 	});
 });
