+++ static/js/jira-manual-app-creation.js
@@ -63,8 +63,6 @@ const gitHubAppPostRequest = (data, token) => {
 		data,
 		success: function() {
 			const child = openChildWindow(`/session/github/${data.uuid}/configuration?ghRedirect=to`);
-			// Remove below line on cleaning up NEW_JWT_VALIDATION flag
-			child.window.jiraHost = jiraHost;
 			child.window.jwt = token;
 		},
 		error: handleFormErrors
