+++ static/js/jira-select-github-cloud-app.js
@@ -29,8 +29,6 @@ $('.select-server').click(function (event) {
   const uuid = $(this).data("identifier");
   window.AP.context.getToken(function(token) {
     const child = openChildWindow(`/session/github/${uuid}/configuration?ghRedirect=to`);
-    // Remove below line on cleaning up NEW_JWT_VALIDATION flag
-    child.window.jiraHost = jiraHost;
     child.window.jwt = token;
   });
 });
